+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- fix structured tool call results and upgrade `rust-mcp-sdk` to `0.7.3` and  ([#3](https://github.com/seaofvoices/rust-mcp-utils/pull/3))
+
 ## 0.1.1
 
 - fix issue related to passing no arguments to tool call ([#1](https://github.com/seaofvoices/rust-mcp-utils/pull/1))